Check hashes of files in download cache for validity
The gomobile init command will use files cached in a 'dl' folder
in lieu of downloading a new file from scratch. If a user of gomobile
does not upgrade for a substantial amount of time, the cached files
will be out of date and possibly cause the init command to fail because
they are missing something the init command now expects.

The init command has a list of expectad sha256 message digests of files
it downloads, which can be utilized to check a file that exists in the 'dl'
folder before proceeeding to utilize it to continue with the init process.

This fixes the problem cited in issue #14503, which is not platform-specific:
https://github.com/golang/go/issues/14503

Some files detailing debugging of the original problem and asserting a solution
are at the following URL:
https://gist.github.com/aoeu/657386c35131b71893153efe34ec07b4